+++ tools/packages/wesl/src/ParseWESL.ts
@@ -4,7 +4,7 @@ import { FlatImport, flattenTreeImport } from "./FlattenTreeImport.ts";
 import { WeslStream } from "./parse/WeslStream.ts";
 import { emptyScope, resetScopeIds, Scope, SrcModule } from "./Scope.ts";
 import { OpenElem } from "./WESLCollect.ts";
-import { weslRoot } from "./parse/WeslGrammar1.ts";
+import { weslRoot } from "./parse/WeslGrammar.ts";
 
 /** result of a parse for one wesl module (e.g. one .wesl file)
  *
